requirements.txt
streamlit
tensorflow
torch
opencv-python
numpy
pillow
matplotlib
scikit-learn
